--- earth low simple.pyr
@@ -1,38 +1,0 @@
-# A demo file for pyrad
-
-setting: low
-
-#         mBar  , K  , km,cm-1,cm-1,  m  , g
-surface: 1013.25, 300, 90, 500, 900, 100, 9.8
-
-#        mol text name, concentration
-molecule: co2, .000400
-molecule: h2o, .018000
-molecule: o3, 0
-molecule: o2, .20
-molecule: n2, .79
-molecule: ch4, .0000018
-
-#           name,      endHeight, endTemp
-lapse rate: troposphere, 11, 216
-lapse rate: tropopause, 20, 216
-lapse rate: stratosphere, 32, 228
-lapse rate: stratosphere, 47, 270
-lapse rate: stratopause, 51, 270
-lapse rate: mesosphere, 71, 214
-lapse rate: mesosphere2, 80, 190
-lapse rate: mesopause, 90, 190
-
-#                 name,           endHeight, endValue, moleculeName
-composition rate: WV boundary layer, 2.5, 18000e-6, h2o
-composition rate: WV troposphere1, 8, 200e-6, h2o
-composition rate: WV troposphere2, 9, 400e-6, h2o
-composition rate: WV troposphere2, 10, 400e-6, h2o
-composition rate: WV tropopause', 20, 2e-6, h2o
-composition rate: WV stratosphere to top',90, 0, h2o
-
-#composition rate: troposphere ozone, 16, 60E-9, o3
-#composition rate: tropopause ozone, 32, 5E-6, o3
-#composition rate: upper strat ozone, 60, 0, o3
-#composition rate: strat on up ozone, 90, 0, o3
-
