+++ earthsimple.pyr
@@ -1,9 +1,9 @@
 # A demo file for pyrad
 
-setting: mid
+setting: low
 
 #         mBar  , K  , km,cm-1,cm-1,  m  , g
-surface: 1013.25, 300, 90, 0, 2000, 100, 9.8
+surface: 1013.25, 300, 120, 100, 1500, 100, 9.8
 
 #        mol text name, concentration
 molecule: co2, .000400
@@ -22,6 +22,7 @@ lapse rate: stratopause, 51, 270
 lapse rate: mesosphere, 71, 214
 lapse rate: mesosphere2, 80, 190
 lapse rate: mesopause, 90, 190
+lapse rate: thermosphere, 100, 228
 
 #                 name,           endHeight, endValue, moleculeName
 composition rate: WV boundary layer, 2.5, 18000e-6, h2o
@@ -31,7 +32,8 @@ composition rate: WV troposphere2, 10, 400e-6, h2o
 composition rate: WV tropopause', 20, 2e-6, h2o
 composition rate: WV stratosphere to top',90, 0, h2o
 
-#composition rate: troposphere ozone, 16, 60E-9, o3
-#composition rate: tropopause ozone, 32, 5E-6, o3
-#composition rate: upper strat ozone, 60, 0, o3
-#composition rate: strat on up ozone, 90, 0, o3
+composition rate: troposphere ozone, 16, 60E-9, o3
+composition rate: tropopause ozone, 32, 5E-6, o3
+composition rate: upper strat ozone, 60, 0, o3
+composition rate: strat on up ozone, 90, 0, o3
+
