--- mars simple.pyr
@@ -1,12 +1,0 @@
-# mars simple
-
-setting: hi
-
-surface: 6, 220, 90, 100, 1500, 200, 3.711
-
-molecule: co2, .9532
-molecule: n2, .027
-molecule: o2, .0013
-
-lapse rate: first layer, 45, 100
-lapse rate: final layer, 90, 100
