+++ marssimple.pyr
@@ -1,0 +1,12 @@
+# mars simple
+
+setting: hi
+
+surface: 6, 220, 100, 100, 1500, 200, 3.711
+
+molecule: co2, .9532
+molecule: n2, .027
+molecule: o2, .0013
+
+lapse rate: first layer, 45, 120
+lapse rate: final layer, 100, 120
