+++ pyrad.py
@@ -1,1387 +1,715 @@
-import os
-import pyradUtilities as utils
-import pyradLineshape as ls
-import pyradIntensity
+import pyradClasses
+import pyradUtilities as util
+import re
 import pyradPlanck
-import numpy as np
-import matplotlib.pyplot as plt
-# import pyradInteractive
+
+existingAtmosphere = False
+validValueAndUnits = re.compile('([-])?(\d+)?([.])?(\d+)?(\S+)?')
+genericAtmosphere = pyradClasses.Atmosphere('holding atm for pyrad interactive')
+pyradClasses.settings.changeSetting('hi')
 
 
-c = 299792458.0
-k = 1.38064852E-23
-h = 6.62607004e-34
-pi = 3.141592653589793
-R = 8.3144598
-t0 = 296
-p0 = 1013.25
-avo = 6.022140857E23
-sb = 5.67E-8
+class Menu:
+    def __init__(self, title, entries, previousMenu=None, menuParams=None):
+        self.title = title
+        self.entries = entries
+        self.previousMenu = previousMenu
+        self.menuParams = menuParams
 
-settings = utils.Settings('low')
+    def displayMenu(self):
+        titleStr = '\t' + self.title
+        while len(titleStr) < 60:
+            titleStr += ' '
+        print('\n%s' % util.underlineCyan(titleStr))
+        i = 1
+        validEntry = ['x']
+        for entry in self.entries:
+            validEntry.append(str(i))
+            print(' %s)   %s' % (util.magentaText(i), entry.name))
+            i += 1
+        if 'Main' not in self.title:
+            print(' %s   Previous menu' % util.magentaText('B)'))
+            validEntry.append('b')
+        print(' %s   Exit' % util.magentaText('X)'))
+        validChoice = False
+        while not validChoice:
+            userInput = input('Choose an option: ')
+            if userInput.lower() == 'x':
+                print('Goodbye')
+                exit(1)
+            elif userInput.lower() == 'b' and 'Main' not in self.title:
+                self.previousMenu()
+            elif userInput in validEntry:
+                userChoice = self.entries[int(userInput) - 1]
+                if userChoice.nextFunction:
+                    userChoice.nextFunction(userChoice.functionParams)
+                    validChoice = True
+                elif userChoice.nextMenu:
+                    userChoice.nextMenu.displayMenu()
+                    validChoice = True
+            else:
+                print('Invalid entry. Try again.')
+
+    def displayMultiChoiceMenu(self):
+        titleStr = '\t' + self.title
+        while len(titleStr) < 60:
+            titleStr += ' '
+        print('\n%s' % util.underlineCyan(titleStr))
+        i = 1
+        validEntry = ['x']
+        for entry in self.entries:
+            validEntry.append(str(i))
+            print(' %s)   %s' % (util.magentaText(i), entry.name))
+            i += 1
+        if 'Main' not in self.title:
+            print(' %s   Previous menu' % util.magentaText('B)'))
+            validEntry.append('b')
+        print(' %s   Exit' % util.magentaText('X)'))
+        validChoice = False
+        while not validChoice:
+            userInput = input('Choose an option: ')
+            if userInput.lower() == 'x':
+                print('Goodbye')
+                exit(1)
+            elif userInput.lower() == 'b' and 'Main' not in self.title:
+                return
+            else:
+                inputs = userInput.split(',')
+                allValid = True
+                userChoices = []
+                for i in inputs:
+                    print('i: %s' % i)
+                    if i.strip() not in validEntry:
+                        allValid = False
+                    else:
+                        userChoices.append(self.entries[int(i) - 1])
+                if allValid:
+                    nextFunction = userChoices[0].nextFunction
+                    nextFunction(userChoices)
+                    validChoice = True
+                else:
+                    print('Invalid entry. Try again.')
 
 
-def reduceRes(array, lumpTo=1):
-    n = int(lumpTo / settings.baseRes)
-    length = int(len(array) * settings.baseRes)
-    newArray = np.zeros(length)
-    for m in range(0, length):
-        for i in range(0, n):
-            newArray[m] += array[m * n + i] / n
-    return newArray
+class Entry:
+    def __init__(self, text, nextMenu=None, nextFunction=None, functionParams=None, previousMenu=None):
+        self.name = text
+        self.nextMenu = nextMenu
+        self.nextFunction = nextFunction
+        self.functionParams = functionParams
+        self.previousMenu = previousMenu
 
 
+def createPlot(params):
+    plotList = params['plots']
+    plotType = params['plotType']
+    plotTitle = params['title']
+    pyradClasses.plot(plotType, plotTitle, plotList)
+    menuChoosePlotType()
 
-def createCustomPlanet(name):
 
-    initialParameters = utils.parseCustomProfile(name)
-    initialValues = initialParameters['initialValues']
-    moleculeList = initialParameters['molecules']
-    temperatureRuleList = initialParameters['temperatureRules']
-    compositionRuleList = initialParameters['compositionRules']
+def plotPlanetSpectrum(values):
+    pList = []
+    for p in values['profiles']:
+        planet = pyradClasses.createCustomPlanet(p.name)
+        pList.append(planet)
+    if values['height'] == -2.71828:
+        pyradClasses.plotPlanetSpectrum(pList, direction=values['direction'], verify=False)
+    else:
+        pyradClasses.plotPlanetSpectrum(pList, direction=values['direction'], height=values['height'], verify=False)
+    atmosphereTransferMenu()
 
+
+def createLayer(atmosphere):
+    defaultLayerName = atmosphere.nextLayerName()
+    getLayerName = '\n%s\n' \
+                   'If no value given, default will be %s : ' \
+                   % (util.underlineCyan('Enter the name of the layer.\t\t'),
+                      util.limeText(defaultLayerName))
+    depth = inputLayerDepth()
+    pressure = inputLayerPressure()
+    temperature = inputLayerTemperature()
+    rangeMin, rangeMax = inputLayerRange()
+    validName = False
+    while not validName:
+        layerName = input(getLayerName)
+        if layerName not in genericAtmosphere:
+            validName = True
+        else:
+            print('Name already taken. Please try again.')
+    layer = atmosphere.addLayer(depth, temperature, pressure, rangeMin, rangeMax, name=layerName)
+    createMolecule(layer)
+    return
+
+
+def createMolecule(layer):
+    addMoleculeLoop = True
+    while addMoleculeLoop:
+        moleculeName = inputMoleculeName()
+        concentration, units = inputMoleculeComposition()
+        tempdict = {units: concentration}
+        molecule = layer.addMolecule(moleculeName, **tempdict)
+        while pyradClasses.totalConcentration(layer) > 1:
+            print("%s total concentration exceeds 100%%" % util.magentaText('***\tWARNING\t***'))
+            menuEditComposition(layer)
+        validInput = False
+        while not validInput:
+            ask = input("Add another molecule to the layer %s " % util.magentaText('(y/n) :'))
+            if ask.strip().lower() == 'y':
+                validInput = True
+            elif ask.strip().lower() == 'n':
+                gasCellMenu()
+
+
+def menuEditLayerParam(layer):
+    editDepth = Entry('Depth', nextFunction=editLayerDepth, functionParams=layer)
+    editRange = Entry('Min or max range', nextFunction=editLayerRange, functionParams=layer)
+    editTemperature = Entry('Temperature',nextFunction=editLayerTemperature, functionParams=layer)
+    editPressure = Entry('Pressure', nextFunction=editLayerPressure, functionParams=layer)
+    entryList = [editDepth, editRange, editTemperature, editPressure]
+    menu = Menu('Choose the parameter to edit for %s' % layer.name, entryList,
+                previousMenu=menuEditParamsOrComp, menuParams=layer)
+    menu.displayMenu()
+    return
+
+
+def editLayerDepth(layer):
+    print('Current %s for %s is : %s\n'
+          % (util.limeText('depth'), util.limeText(layer.name), util.cyanText('%scm' % layer.depth)))
+    depth = inputLayerDepth(default=layer.depth)
+    layer.changeDepth(depth)
+    menuEditLayerParam(layer)
+
+
+def editLayerTemperature(layer):
+    print('Current %s for %s is : %s\n'
+          % (util.limeText('temperature'), util.limeText(layer.name), util.cyanText('%sK' % layer.T)))
+    temperature = inputLayerTemperature(default=layer.T)
+    layer.changeTemperature(temperature)
+    menuEditLayerParam(layer)
+
+
+def editLayerPressure(layer):
+    print('Current %s for %s is : %s\n'
+          % (util.limeText('pressure'), util.limeText(layer.name), util.cyanText('%smbar' % layer.P)))
+    pressure = inputLayerPressure(default=layer.P)
+    layer.changePressure(pressure)
+    menuEditLayerParam(layer)
+
+
+def editLayerRange(layer):
+    print('Current %s for %s is %s\n'
+          % (util.limeText('range'), util.limeText(layer.name),
+             util.cyanText('%s-%scm-1' % (layer.rangeMin, layer.rangeMax))))
+    rangeMin, rangeMax = inputLayerRange(defaultMin=layer.rangeMin, defaultMax=layer.rangeMax)
+    layer.changeRange(rangeMin, rangeMax)
+    menuEditLayerParam(layer)
+
+
+def editComposition(molecule):
+    print('Current concentration for %s is %s\n' % (util.limeText(molecule.name), util.limeText(molecule.concText)))
+    return inputMoleculeComposition(molecule, default=molecule.concText)
+
+
+def inputLayerDepth(default=None):
+    if not default:
+        default = 10
+    text = 'Enter the thickness of the layer.\t\t\t'
+    getDepth = '%s\n' \
+               'If no units are specified, %s will be assumed.\n' \
+               'Other valid units are %s . If no value given, default will be %s:  ' % \
+        (util.underlineCyan(text),
+         util.limeText('cm'),
+         util.limeText('m, in, ft.'),
+         util.limeText('%scm' % default))
+    depth = receiveInput(getDepth, validDepth, default=default)
+    return depth
+
+
+def inputLayerTemperature(default=None):
+    if not default:
+        default = 300
+    text = 'Enter the temperature of the layer.\t\t\t'
+    getTemperature = '%s\n' \
+                     'If no units are specified, %s will be assumed.\n' \
+                     'Other valid units are %s . If no value given, default will be %s:  ' % \
+                    (util.underlineCyan(text),
+                     util.limeText('K'),
+                     util.limeText('C or F'),
+                     util.limeText('%sK' % default))
+    temperature = receiveInput(getTemperature, validTemperature, default=default)
+    return temperature
+
+
+def inputLayerPressure(default=None):
+    if not default:
+        default = 1013.25
+    text = 'Enter the pressure of the layer.\t\t\t'
+    getPressure = '%s\n' \
+                  'If no units are specified, %s will be assumed.\n' \
+                  'Other valid units are %s . If no value given, default will be %s:  ' % \
+                  (util.underlineCyan(text),
+                   util.limeText('mBar'),
+                   util.limeText('pa, bar, and atm.'),
+                   util.limeText('%smbar' % default))
+    pressure = receiveInput(getPressure, validPressure, default=default)
+    return pressure
+
+
+def inputLayerRange(defaultMin=None, defaultMax=None):
+    if not defaultMin:
+        defaultMin = 600
+    if not defaultMax:
+        defaultMax = 700
+    rangeMin = -1
+    rangeMax = -1
+    text = 'Enter the minimum range of the layer.\t\t\t'
+    getRangeMin = '%s\n' \
+                  'If no units are specified, %s will be assumed.\n' \
+                  'Other valid units are %s . If no value given, default will be %s:  ' % \
+                  (util.underlineCyan(text),
+                   util.limeText('cm-1'),
+                   util.limeText('um'),
+                   util.limeText('%scm' % defaultMin))
+    while rangeMin < 0:
+        rangeMin = receiveInput(getRangeMin, validRange, default=defaultMin)
+        if rangeMin < 0:
+            print('Range min must be %s than zero' % util.magentaText('greater'))
+    text = 'Enter the maximum range of the layer.\t\t\t'
+    getRangeMax = '%s\n' \
+                  'If no units are specified, %s will be assumed.\n' \
+                  'Other valid units are %s . If no value given, default will be %s:  ' % \
+                  (util.underlineCyan(text),
+                   util.limeText('cm-1'),
+                   util.limeText('um'),
+                   util.limeText('%scm' % defaultMax))
+    while rangeMax <= rangeMin:
+        rangeMax = receiveInput(getRangeMax, validRange, default=defaultMax)
+        if rangeMax <= rangeMin:
+            print('Range min must be %s than range min of %s' % (util.magentaText('greater'), util.cyanText(rangeMin)))
+    return rangeMin, rangeMax
+
+
+def validNumber(userInput):
     try:
-        value = initialParameters['setting']
-    except KeyError:
-        value = 'low'
+        return float(userInput)
+    except ValueError:
+        return False
 
-    settings.changeSetting(value)
-    planet = Planet(initialParameters['name'], initialValues['surfacePressure'], initialValues['surfaceTemperature'],
-                    initialValues['maxHeight'], rangeMin=initialValues['rangeMin'],
-                    rangeMax=initialValues['rangeMax'], initialThickness=initialValues['initialDepth'], gravity=initialValues['gravity'])
 
+def inputPlanckRange(units):
+    rangeMin = -1
+    rangeMax = -1
+    text = '%s\nUnits are %s. Scientific notation is accepted (1e14):' \
+           % (util.underlineCyan('Enter the minimum range of the planck spectrum.'),util.limeText(units))
+    while rangeMin < 0:
+        rangeMin = receiveInput(text, validNumber)
+        if rangeMin < 0:
+            print('Range min must be %s than zero' % util.magentaText('greater'))
+    text = '%s\nUnits are %s. Scientific notation is accepted (1e14):' \
+           % (util.underlineCyan('Enter the maximum range of the planck spectrum.'), util.limeText(units))
+    while rangeMax <= rangeMin:
+        rangeMax = receiveInput(text, validNumber)
+        if rangeMax <= rangeMin:
+            print('Range min must be %s than range min of %s' % (util.magentaText('greater'), util.cyanText(rangeMin)))
+    return rangeMin, rangeMax
+
+
+def inputMoleculeName(default=None):
+    if not default:
+        default = 'co2'
+    text = 'Enter the short molecule name.\t\t\t'
+    moleculeName = receiveInput('%s\n'
+                                'For a full list of options, type %s . If no value given, %s will be used: '
+                                % (util.underlineCyan(text),
+                                   util.magentaText('help'),
+                                   util.limeText(default)), validMoleculeName, default=default)
+    return moleculeName
+
+
+def inputMoleculeComposition(obj=None, default=None):
+    if not default:
+        default = '400ppm'
+    text = 'Enter the molecule composition.\t\t\t'
+    composition, units = receiveInput('%s\n'
+                                      'If no units entered, composition will be assumed %s.\n'
+                                      'Other valid units are %s . If no value given, %s will be used: '
+                                      % (util.underlineCyan(text),
+                                         util.limeText('parts per 1'),
+                                         util.limeText('ppm, ppb, or percentage'),
+                                         util.limeText(default)), validComposition, default=default)
+    if obj:
+        if units == 'ppm':
+            obj.setPPM(composition)
+        elif units == 'ppb':
+            obj.setPPB(composition)
+        elif 'perc' in units or units == '%':
+            obj.setPercentage(composition)
+        else:
+            obj.setConcentrationPercentage(composition)
+        return
+    else:
+        return composition, units
+
+
+def inputPlanckTemps():
+    text = 'Enter the temperature of the planck curves.\t\t\t'
+    tempList = receiveMultiInput('%s\n'
+                            'If no units entered, temperature will be assumed %s.\n'
+                            'Other valid units are %s .Multiple temperatures can be separated with a comma: '
+                            % (util.underlineCyan(text),
+                               util.limeText('K'),
+                               util.limeText('C and F')), validTemperature)
+    return tempList
+
+
+def inputHeight(values):
+    text = 'Enter the height to view transmission from.\t\t\t'
+    height = receiveInput('%s\n'
+                          'Units should be in %s. If no value entered, maximum atm height will be used: ' % (util.underlineCyan(text), util.limeText('km')), validNumber, default=-2.71828)
+    values['height'] = height
+    plotPlanetSpectrum(values)
+    return
+
+
+def menuChooseLayerToEdit(empty=None):
+    entryList = []
+    for layer in genericAtmosphere:
+        nextEntry = Entry(layer.name, nextFunction=menuEditParamsOrComp, functionParams=layer)
+        entryList.append(nextEntry)
+    editLayerMenu = Menu('Edit layer', entryList)
+    editLayerMenu.displayMenu()
+    return
+
+
+def menuChooseTransmission(plotType):
+    entryList = []
+    for layer in genericAtmosphere:
+        params = {'plots': [layer], 'plotType': plotType, 'title': layer.title}
+        nextEntry = Entry(layer.name, nextFunction=createTransmission, functionParams=params)
+        entryList.append(nextEntry)
+        plotList, title = createObjAndComponents(layer)
+        params = {'plots': plotList, 'plotType': plotType, 'title': title}
+        nextEntry = Entry('%s and components' % layer.name, nextFunction=createTransmission, functionParams=params)
+        entryList.append(nextEntry)
+    transmissionMenu = Menu('Choose which layers to plot transmission', entryList)
+    transmissionMenu.displayMenu()
+    return
+
+
+def createPlanckCurves(plotType):
+    plotList = inputPlanckTemps()
+    rangeMin, rangeMax = inputPlanckRange(plotType)
+    pyradClasses.plotSpectrum(title='Planck spectrums', rangeMin=rangeMin, rangeMax=rangeMax, planckTemperatureList=plotList, planckType=plotType)
+    return
+
+
+def menuPlanckType(empty=None):
+    entryList = []
+    wavenumber = 'wavenumber'
+    hertz = 'Hz'
+    wavelength = 'wavelength'
+    entryList.append(Entry('By %s (cm-1)' % wavenumber, nextFunction=createPlanckCurves, functionParams=wavenumber))
+    entryList.append(Entry('By %s (um)' % wavelength, nextFunction=createPlanckCurves, functionParams=wavelength))
+    entryList.append(Entry('By %s (s-1)' % hertz, nextFunction=createPlanckCurves, functionParams=hertz))
+    planckTypeMenu = Menu('Choose planck type', entryList)
+    planckTypeMenu.displayMenu()
+    return
+
+
+def createTransmission(params):
+    layer = params['plots'][0]
+    text = 'A plot for transmission requires an initial surface temperature.\n'\
+           'Please choose a temperature different from the layer temperature of %sK:' % util.limeText(layer.T)
+    temperature = receiveMultiInput(text, validTemperature)
+    objList = []
+    for item in params['plots']:
+        objList.append(item)
+    temperature.append(layer.T)
+    pyradClasses.plotSpectrum(layer, objList=objList,
+                              surfaceSpectrum=pyradPlanck.planckWavenumber(layer.xAxis, temperature[0]),
+                              planckTemperatureList=temperature)
+    return
+
+
+def menuChoosePlotType(empty=None):
+    entryList = []
+    entryList.append(Entry('transmittance', nextFunction=menuChooseLayerToPlot, functionParams='transmittance'))
+    entryList.append(Entry('absorption coefficient', nextFunction=menuChooseLayerToPlot, functionParams='absorption coefficient'))
+    entryList.append(Entry('cross section', nextFunction=menuChooseLayerToPlot, functionParams='cross section'))
+    entryList.append(Entry('absorbance', nextFunction=menuChooseLayerToPlot, functionParams='absorbance'))
+    entryList.append(Entry('optical depth', nextFunction=menuChooseLayerToPlot, functionParams='optical depth'))
+    entryList.append(Entry('line survey', nextFunction=menuChooseLayerToPlot, functionParams='line survey'))
+    entryList.append(Entry('transmission', nextFunction=menuChooseTransmission, functionParams='transmission'))
+    choosePlotTypeMenu = Menu('Choose plot type', entryList, previousMenu=gasCellMenu)
+    choosePlotTypeMenu.displayMenu()
+    return
+
+
+def menuChooseLayerToPlot(plotType):
+    entryList = []
+    for layer in genericAtmosphere:
+        params = {'plots': [layer], 'plotType': plotType, 'title': layer.title}
+        nextEntry = Entry(layer.name, nextFunction=createPlot, functionParams=params)
+        entryList.append(nextEntry)
+        plotList, title = createObjAndComponents(layer)
+        params = {'plots': plotList, 'plotType': plotType, 'title': title}
+        nextEntry = Entry('%s and components' % layer.name, nextFunction=createPlot, functionParams=params)
+        entryList.append(nextEntry)
+    plotLayerMenu = Menu('Plot layer', entryList, previousMenu=menuChoosePlotType)
+    plotLayerMenu.displayMenu()
+    return
+
+
+def createObjAndComponents(obj):
+    plotList = [obj]
+    for item in obj:
+        plotList.append(item)
+    return plotList, obj.title
+
+
+def menuEditComposition(layer):
+    moleculeList = layer.returnMoleculeObjects()
+    entryList = []
     for molecule in moleculeList:
-        planet.addMolecule(molecule['name'], concentration=molecule['concentration'])
-
-    for rule in temperatureRuleList:
-        planet.addLapseRate(rule['name'], rule['finalHeight'], rule['finalValue'])
-
-    for rule in compositionRuleList:
-        initalLayer = planet.initialLayer
-        molecule = initalLayer.returnMolecule(rule['moleculeName'])
-        planet.addCompositionRate(rule['name'], rule['finalHeight'], rule['finalValue'], molecule)
-    return planet
+        newEntry = Entry('%s : %s' % (molecule.name, molecule.concText),
+                         functionParams=molecule, nextFunction=inputMoleculeComposition)
+        entryList.append(newEntry)
+    entryList.append(Entry('Add a new molecule(s)', nextFunction=createMolecule, functionParams=layer))
+    editCompMenu = Menu('Choose a molecule to edit', entryList, previousMenu=menuEditParamsOrComp)
+    editCompMenu.displayMenu()
+    return
 
 
-def yesOrNo(promptText):
+def menuEditParamsOrComp(layer):
+    entryList = []
+    editLayerParamsEntry = Entry('Edit layer parameters', nextFunction=menuEditLayerParam, functionParams=layer)
+    entryList.append(editLayerParamsEntry)
+    duplicateLayerEntry = Entry('Duplicate layer', nextFunction=duplicateObj, functionParams=layer)
+    entryList.append(duplicateLayerEntry)
+    editCompositionEntry = Entry('Edit composition', nextFunction=menuEditComposition, functionParams=layer)
+    entryList.append(editCompositionEntry)
+    chooseParamsMenu = Menu('Edit or duplicate', entryList, previousMenu=menuChooseLayerToEdit)
+    chooseParamsMenu.displayMenu()
+    return
+
+
+def gasCellMenu(param=None):
+    createLayerEntry = Entry("Create new gas cell", nextFunction=createLayer, functionParams=genericAtmosphere)
+    editLayerEntry = Entry("Edit/duplicate gas cell", nextFunction=menuChooseLayerToEdit)
+    plotLayerEntry = Entry("Plot gas cell", nextFunction=menuChoosePlotType)
+    planckPlotEntry = (Entry('Plot planck curves', nextFunction=menuPlanckType))
+    menuGasCell = Menu('Gas cell simulator', [createLayerEntry, editLayerEntry, plotLayerEntry, planckPlotEntry], previousMenu=menuMain)
+    menuGasCell.displayMenu()
+    return
+
+
+def chooseDirection(profileList):
+    lookUpEntry = Entry('Looking up', nextFunction=inputHeight, functionParams={'profiles': profileList,
+                                                                                'direction': 'up'})
+    lookDownEntry = Entry('Looking down', nextFunction=inputHeight, functionParams={'profiles': profileList,
+                                                                                'direction': 'down'})
+    menuChooseDirection = Menu('Choose direction to look', [lookUpEntry, lookDownEntry], previousMenu=atmosphereTransferMenu)
+    menuChooseDirection.displayMenu()
+    return
+
+
+def atmosphereTransferMenu(param=None):
+    profileList = util.getProfileList()
+    entryList = []
+    for profile in profileList:
+        entryList.append(Entry('%s' % profile[:-4], nextFunction=chooseDirection, functionParams=profile))
+    menuAtmTransfer = Menu('Choose atmosphere', entryList, previousMenu=menuMain)
+    menuAtmTransfer.displayMultiChoiceMenu()
+    return
+
+
+def menuMain():
+    gasCellEntry = Entry('Gas cell simulator', nextFunction=gasCellMenu)
+    atmosphereTransferEntry = Entry('Atmosphere transmission', nextFunction=atmosphereTransferMenu)
+    mainMenu = Menu('Main menu', [gasCellEntry, atmosphereTransferEntry])
+    mainMenu.displayMenu()
+    return
+
+
+def duplicateObj(obj):
+    newObj = obj.returnCopy()
+    if isinstance(newObj, pyradClasses.Layer):
+        genericAtmosphere.append(newObj)
+    else:
+        print('Unknown object %s, type %s' % (obj.name, type(obj)))
+    return
+
+
+def receiveInput(inputText, validInputFunction, default=None):
+    validInput = False
+    while validInput is False:
+        userInput = input('\n%s' % inputText)
+        if userInput == '':
+            return validInputFunction(str(default))
+        validInput = validInputFunction(userInput)
+    return validInput
+
+
+def receiveMultiInput(inputText, validInputFunction, default=None):
     validInput = False
     while not validInput:
-        userSelection = input(promptText)
-        if not userSelection:
-            pass
-        elif userSelection.lower()[0] == 'y':
-            return True
-        elif userSelection.lower()[0] == 'n':
-            return False
-        else:
-            print('Invalid choice.')
-
-
-def smooth(y, box_pts):
-    box = np.ones(box_pts)/box_pts
-    y_smooth = np.convolve(y, box, mode='same')
-    return y_smooth
-
-
-def linear(baseValue, baseHeight, rate, height):
-    newValue = baseValue + rate * (height - baseHeight)
-    return newValue
-
-
-def integrateSpectrum(spectrum, unitAngle=pi, res=utils.BASE_RESOLUTION):
-    value = np.sum(np.nan_to_num(spectrum))
-    value = value * unitAngle * res
-    return value
-
-
-def getCrossSection(obj):
-    if not obj.progressCrossSection:
-        obj.createCrossSection()
-    return obj.crossSection
-
-
-def resetCrossSection(obj):
-    if not obj.progressCrossSection:
-        return
-    obj.crossSection = np.zeros(int((obj.rangeMax - obj.rangeMin) / utils.BASE_RESOLUTION))
-    obj.progressCrossSection = False
-    for child in obj:
-        if not isinstance(child, Line):
-            resetCrossSection(child)
-
-
-def resetData(obj):
-    # clears the existing line data from parent object down to isotope, and then reloads the data using getData
-    # use this if layer ranges get changed. Will also clear the cross section data of the obj.
-    for child in obj:
-        if isinstance(child, Isotope):
-            while len(child) > 0:
-                child.pop()
-            child.getData()
-        else:
-            resetData(child)
-    resetCrossSection(obj)
-
-
-def getAbsCoef(obj):
-    if not obj.progressCrossSection:
-        obj.createCrossSection()
-    return obj.absCoef
-
-
-def getTransmittance(obj):
-    if not obj.progressCrossSection:
-        obj.createCrossSection()
-    return obj.transmittance
-
-
-def getOpticalDepth(obj):
-    if not obj.progressCrossSection:
-        obj.createCrossSection()
-    return -np.log(obj.transmittance)
-
-
-def getAbsorbance(obj):
-    if not obj.progressCrossSection:
-        obj.createCrossSection()
-    return obj.absorbance
-
-
-def getEmissivity(obj):
-    if not obj.progressCrossSection:
-        obj.createCrossSection()
-    return obj.emissivity
-
-
-def getGlobalIsotope(ID, isotopeDepth):
-    globalIsoList = []
-    for i in range(1, isotopeDepth + 1):
-        globalIsoList.append(HITRAN_GLOBAL_ISO[ID][i])
-    return globalIsoList
-
-
-def totalConcentration(layer):
-    total = 0
-    for molecule in layer:
-        total += molecule.concentration
-    return total
-
-
-def totalLineList(obj):
-    fullList = []
-    if isinstance(obj, Isotope):
-        return obj.linelist()
-    for item in obj:
-        fullList += totalLineList(item)
-    return fullList
-
-
-def convertLength(value, units):
-    if units == 'cm':
-        return value
-    elif units in ['m', 'meter']:
-        return value * 100
-    elif units in ['ft', 'feet']:
-        return value * 30.48
-    elif units in ['in', 'inch']:
-        return value * 2.54
-
-
-def convertPressure(value, units):
-    if units == 'mbar':
-        return value
-    elif units in ['atm', 'atmospheres', 'atmosphere']:
-        return value * 1013.25
-    elif units in ['b', 'bar']:
-        return value * 1000
-    elif units in ['pa', 'pascal', 'pascals']:
-        return value / 100
-
-
-def convertRange(value, units):
-    if units == 'cm-1':
-        return value
-    elif units in ['um', 'micrometers', 'micrometer']:
-        return 10000 / value
-
-
-def convertTemperature(value, units):
-    if units[0].upper() == 'K':
-        return value
-    elif units[0].upper() == 'C':
-        return value + 273
-    elif units[0].upper() == 'F':
-        return (value - 32) * 5 / 9 + 273
-
-
-def interpolateArray(hiResXAxis, loResXAxis, loResYValues):
-    hiResY = np.interp(hiResXAxis, loResXAxis, loResYValues)
-    return hiResY
-
-
-class Line:
-    def __init__(self, wavenumber, intensity, einsteinA, airHalfWidth,
-                 selfHalfWidth, lowerEnergy, tempExponent, pressureShift, parent):
-        self.isotope = parent
-        self.molecule = self.isotope.molecule
-        self.layer = self.molecule.layer
-        self.wavenumber = wavenumber
-        self.intensity = intensity
-        self.einsteinA = einsteinA
-        self.airHalfWidth = airHalfWidth
-        self.selfHalfWidth = selfHalfWidth
-        self.lowerEnergy = lowerEnergy
-        self.tempExponent = tempExponent
-        self.pressureShift = pressureShift
-
-    @property
-    def broadenedLine(self):
-        return self.wavenumber + self.pressureShift * self.layer.P / p0
-
-    @property
-    def lorentzHW(self):
-        return (float((1 - self.molecule.concentration) * self.airHalfWidth + self.molecule.concentration
-                * self.selfHalfWidth) * (self.layer.P / p0) * (t0 / self.layer.T) ** self.tempExponent)
-
-    @property
-    def gaussianHW(self):
-        return self.broadenedLine * np.sqrt(2 * k * self.layer.T / self.isotope.molMass / c ** 2)
-
-
-class Isotope(list):
-    def __init__(self, number, molecule):
-        super(Isotope, self).__init__(self)
-        params = utils.readMolParams(number)
-        self.globalIsoNumber = params[0]
-        self.shortName = params[1]
-        self.name = 'Isotope %s' % self.globalIsoNumber
-        self.molNum = params[2]
-        self.isoN = params[3]
-        self.abundance = params[4]
-        self.q296 = params[5]
-        self.gj = params[6]
-        self.molmass = params[7]
-        self.molecule = molecule
-        self.layer = self.molecule.layer
-        self.q = {}
-        self.crossSection = np.copy(self.layer.crossSection)
-        self.lineSurvey = np.zeros(int((self.layer.rangeMax - self.layer.rangeMin) / utils.BASE_RESOLUTION))
-        self.progressCrossSection = False
-
-    @property
-    def P(self):
-        return self.layer.P
-
-    @property
-    def molMass(self):
-        return self.molmass / 1000 / avo
-
-    @property
-    def T(self):
-        return self.layer.T
-
-    @property
-    def depth(self):
-        return self.layer.depth
-
-    @property
-    def rangeMin(self):
-        return self.layer.rangeMin
-
-    @property
-    def rangeMax(self):
-        return self.layer.rangeMax
-
-    @property
-    def resolution(self):
-        return self.layer.resolution
-
-    @property
-    def distanceFromCenter(self):
-        return self.layer.distanceFromCenter
-
-    @property
-    def absCoef(self):
-        return self.crossSection * self.molecule.concentration * self.layer.P / 1E4 / k / self.layer.T
-
-    @property
-    def transmittance(self):
-        return np.exp(-self.absCoef * self.layer.depth)
-
-    @property
-    def emissivity(self):
-        return 1 - self.transmittance
-
-    @property
-    def emittance(self):
-        return self.emissivity
-
-    @property
-    def absorbance(self):
-        return np.log10(1 / self.transmittance)
-
-    @property
-    def yAxis(self):
-        return np.copy(self.layer.yAxis)
-
-    @property
-    def xAxis(self):
-        return np.copy(self.layer.xAxis)
-
-    def getData(self, lineSurvey=False, verbose=True):
-        if verbose:
-            print('Getting data for %s, %s, isotope %s' % (self.layer.name, self.molecule.name, self.globalIsoNumber), end='\r', flush=True)
-        lineDict = utils.gatherData(self.globalIsoNumber, self.layer.effectiveRangeMin, self.layer.effectiveRangeMax)
-        self.q = utils.getQData(self.globalIsoNumber)
-        for line in lineDict:
-            if lineDict[line]['intensity'] > settings.lineIntensityCutoff:
-                self.append(Line(line, lineDict[line]['intensity'], lineDict[line]['einsteinA'],
-                             lineDict[line]['airHalfWidth'], lineDict[line]['selfHalfWidth'],
-                             lineDict[line]['lowerEnergy'], lineDict[line]['tempExponent'],
-                             lineDict[line]['pressureShift'], self))
-        # self.createLineSurvey()
-
-    def createCrossSection(self):
-        molecule = self.molecule
-        layer = molecule.layer
-        progress = 0
-        i = 1
-        alertInterval = int(len(self) / 20)
-        crossSection = np.copy(self.yAxis)
-        trackGauss = 0
-        trackLorentz = 0
-        trackVoigt = 0
-        if len(self) == 0:
-            self.progressCrossSection = True
-            return
-        for line in self:
-            if progress > i * alertInterval and len(self) > 50:
-                text = 'Processing %s, height %skm, mean height %skm, molecule %s: <%s%s>' % (layer.name, round(layer.height / 100000, 2), round(layer.meanHeight / 100000, 2),
-                         molecule.name, '*' * i, '-' * (20 - i))
-                print(text, end='\r', flush=True)
-                i += 1
-            progress += 1
-            xValues = np.arange(0, layer.distanceFromCenter, layer.resolution)
-            hwRatio = line.lorentzHW / line.gaussianHW
-            if hwRatio < .01:
-                rightCurve = ls.gaussianLineShape(line.gaussianHW, xValues)
-                trackGauss += 1
-            elif hwRatio > 100:
-                rightCurve = ls.lorentzLineShape(line.lorentzHW, xValues)
-                trackLorentz += 1
-            else:
-                rightCurve = ls.pseudoVoigtShape(line.gaussianHW, line.lorentzHW, xValues)
-                trackVoigt += 1
-            intensity = pyradIntensity.intensityFactor(line.intensity, line.broadenedLine,
-                                                       layer.T, line.lowerEnergy, self.q[layer.T], self.q296)
-            arrayIndex = int((line.wavenumber - layer.rangeMin) / layer.resolution)
-            arrayLength = len(crossSection) - 1
-            if isBetween(arrayIndex, 0, arrayLength):
-                crossSection[arrayIndex] = crossSection[arrayIndex] + rightCurve[0] * intensity
-            for dx in range(1, len(rightCurve) - 1):
-                rightIndex = arrayIndex + dx
-                leftIndex = arrayIndex - dx
-                if isBetween(rightIndex, 0, arrayLength):
-                    crossSection[rightIndex] += rightCurve[dx] * intensity
-                if isBetween(leftIndex, 0, arrayLength):
-                    crossSection[leftIndex] += rightCurve[dx] * intensity
-        self.crossSection = interpolateArray(self.xAxis,
-                                             np.linspace(self.rangeMin, self.rangeMax,
-                                                         (self.rangeMax - self.rangeMin) / self.resolution,
-                                                         endpoint=True),
-                                             crossSection)
-        self.progressCrossSection = True
-
-    def createLineSurvey(self):
-        print('Creating line survey for %s' % self.name)
-        molecule = self.molecule
-        layer = molecule.layer
-        progress = 0
-        i = 1
-        alertInterval = int(len(self) / 20)
-        lineSurvey = np.zeros(int((self.rangeMax - self.rangeMin) / utils.BASE_RESOLUTION))
-        for line in self:
-            if progress > i * alertInterval:
-                print('Progress for %s <%s%s>' % (molecule.name, '*' * i, '-' * (20 - i)), end='\r', flush=True)
-                i += 1
-            progress += 1
-            intensity = line.intensity
-            arrayIndex = int((line.wavenumber - layer.rangeMin) / layer.resolution)
-            arrayLength = len(lineSurvey) - 1
-            if isBetween(arrayIndex, 0, arrayLength):
-                lineSurvey[arrayIndex] = lineSurvey[arrayIndex] + intensity
-        self.lineSurvey = lineSurvey
-        return self.lineSurvey
-
-    def linelist(self):
-        lines = []
-        for line in self:
-            lines.append(line)
-        return lines
-
-    def planck(self, temperature):
-        return self.layer.planck(temperature)
-
-    def transmission(self, surfaceSpectrum):
-        transmitted = self.transmittance * surfaceSpectrum
-        emitted = self.emittance * self.planck(self.T)
-        return transmitted + emitted
-
-    def returnCopy(self):
-        return self.linelist()
-
-
-class Molecule(list):
-    def __init__(self, shortNameOrMolNum, layer, isotopeDepth=1, **abundance):
-        super(Molecule, self).__init__(self)
-        self.layer = layer
-        self.crossSection = np.copy(layer.crossSection)
-        self.isotopeDepth = isotopeDepth
-        self.concText = ''
-        self.concentration = 0
-        try:
-            int(shortNameOrMolNum)
-            self.ID = int(shortNameOrMolNum)
-            self.name = False
-        except ValueError:
-            self.name = shortNameOrMolNum
-            self.ID = MOLECULE_ID[self.name]
-        for isotope in getGlobalIsotope(self.ID, isotopeDepth):
-            isoClass = Isotope(isotope, self)
-            self.append(isoClass)
-            if not self.name:
-                self.name = isoClass.shortName
-        self.progressCrossSection = False
-        for key in abundance:
-            if key == 'ppm':
-                self.setPPM(abundance[key])
-            elif key == 'ppb':
-                self.setPPB(abundance[key])
-            elif key == 'percentage' or key == 'perc' or key == '%':
-                self.setPercentage(abundance[key])
-            elif key == 'concentration':
-                self.setConcentration(abundance[key])
-            else:
-                print('Invalid concentration type. Use ppm, ppb, percentage, or concentration.')
-
-    def __str__(self):
-        return '%s: %s' % (self.name, self.concText)
-
-    def __bool__(self):
-        return True
-
-    def returnCopy(self):
-        valueUnit = self.concText.split()
-        tempDict = {valueUnit[1]: float(valueUnit[0])}
-        newMolecule = Molecule(self.name, self.layer, isotopeDepth=int(self.isotopeDepth), **tempDict)
-        newMolecule.getData(verbose=False)
-        return newMolecule
-
-    def setPercentage(self, percentage):
-        self.concentration = percentage / 100
-        self.concText = '%s %%' % percentage
-        resetCrossSection(self)
-
-    def setPPM(self, ppm):
-        self.concentration = ppm * 10**-6
-        self.concText = '%s ppm' % ppm
-        resetCrossSection(self)
-
-    def setPPB(self, ppb):
-        self.concentration = ppb * 10**-8
-        self.concText = '%s ppb' % ppb
-        resetCrossSection(self)
-
-    def setConcentration(self, concentration):
-        self.setPPM(concentration * 1E6)
-        resetCrossSection(self)
-
-    def getData(self, verbose=True):
-        for isotope in self:
-            isotope.getData(verbose)
-
-    def createCrossSection(self):
-        tempAxis = np.zeros(int((self.rangeMax - self.rangeMin) / utils.BASE_RESOLUTION))
-        for isotope in self:
-            tempAxis += getCrossSection(isotope)
-        self.progressCrossSection = True
-        self.crossSection = tempAxis
-
-    def planck(self, temperature):
-        return self.layer.planck(temperature)
-
-    def transmission(self, surfaceSpectrum):
-        transmitted = self.transmittance * surfaceSpectrum
-        emitted = self.emittance * self.planck(self.T)
-        return transmitted + emitted
-
-    @property
-    def absCoef(self):
-        return self.crossSection * self.concentration * self.layer.P / 1E4 / k / self.layer.T
-
-    @property
-    def transmittance(self):
-        return np.exp(-self.absCoef * self.layer.depth)
-
-    @property
-    def lineSurvey(self):
-        tempAxis = np.zeros(int((self.rangeMax - self.rangeMin) / utils.BASE_RESOLUTION))
-        for isotope in self:
-            tempAxis += isotope.lineSurvey
-        return tempAxis
-
-    @property
-    def absorbance(self):
-        return np.log10(1 / self.transmittance)
-
-    @property
-    def emissivity(self):
-        return 1 - self.transmittance
-
-    @property
-    def emittance(self):
-        return self.emissivity
-
-    @property
-    def P(self):
-        return self.layer.P
-
-    @property
-    def yAxis(self):
-        return np.copy(self.layer.yAxis)
-
-    @property
-    def xAxis(self):
-        return np.copy(self.layer.xAxis)
-
-    @property
-    def T(self):
-        return self.layer.T
-
-    @property
-    def depth(self):
-        return self.layer.depth
-
-    @property
-    def rangeMin(self):
-        return self.layer.rangeMin
-
-    @property
-    def rangeMax(self):
-        return self.layer.rangeMax
-
-    @property
-    def resolution(self):
-        return self.layer.resolution
-
-    @property
-    def distanceFromCenter(self):
-        return self.layer.distanceFromCenter
-
-    @property
-    def molarMass(self):
-        return self[0].molmass
-
-
-class Layer(list):
-    hasAtmosphere = False
-
-    def __init__(self, depth, T, P, rangeMin, rangeMax, height=0.0, atmosphere=None, name='', dynamicResolution=True):
-        super(Layer, self).__init__(self)
-        self.rangeMin = rangeMin
-        self.rangeMax = rangeMax
-        self.T = T
-        self.P = P
-        self.depth = depth
-        self.distanceFromCenter = self.P / 1013.25 * 5
-        self.effectiveRangeMin = max(self.rangeMin - self.distanceFromCenter, 0)
-        self.effectiveRangeMax = self.rangeMax + self.distanceFromCenter
-        self.dynamicResolution = dynamicResolution
-        self.surfaceSpectrum = None
-        self.height = height
-        if not dynamicResolution:
-            self.resolution = utils.BASE_RESOLUTION
-        else:
-            self.resolution = max(10**int(np.log10((self.P / 1013.25))) * .01, utils.BASE_RESOLUTION)
-        if not atmosphere:
-            if not Layer.hasAtmosphere:
-                self.atmosphere = Atmosphere('generic')
-                Layer.hasAtmosphere = self.atmosphere
-            else:
-                self.atmosphere = Layer.hasAtmosphere
-        else:
-            self.atmosphere = atmosphere
-            self.hasAtmosphere = atmosphere
-        self.crossSection = np.zeros(int((rangeMax - rangeMin) / utils.BASE_RESOLUTION))
-        self.absorptionCoefficient = np.zeros(int((rangeMax - rangeMin) / utils.BASE_RESOLUTION))
-        self.progressCrossSection = False
-        self.progressAbsCoef = False
-        if not name:
-            name = '%s' % self.atmosphere.nextLayerName()
-        self.atmosphere.append(self)
-        self.name = name
-
-    def __str__(self):
-        return '%s; %s' % (self.name, '; '.join(str(m) for m in self))
-
-    def __bool__(self):
-        return True
-
-    def createCrossSection(self):
-        tempAxis = np.zeros(int((self.rangeMax - self.rangeMin) / utils.BASE_RESOLUTION))
-        for molecule in self:
-            tempAxis += getCrossSection(molecule)
-        self.progressCrossSection = True
-        self.crossSection = tempAxis
-
-    def returnMolecule(self, name):
-        for m in self:
-            if m.name == name:
-                return m
+        userInput = input(inputText)
+        inputList = userInput.replace(' ', '').split(',')
+        testInput = True
+        for item in inputList:
+            if not validInputFunction(item):
+                print('%s not recognized as valid. Please try again.')
+                testInput = False
+        if testInput:
+            return inputList
+
+
+def validMoleculeName(userInput):
+    if not userInput:
+        return False
+    if userInput.strip().lower() == 'help':
+        util.displayAllMolecules()
+        return False
+    elif userInput in pyradClasses.MOLECULE_ID:
+        return userInput
+    else:
+        print('Invalid molecule name. %s' % (util.underlineMagenta('Please try again.')))
         return False
 
-    @property
-    def meanHeight(self):
-        return self.height + .5 * self.depth
 
-    @property
-    def lineSurvey(self):
-        tempAxis = np.zeros(int((self.rangeMax - self.rangeMin) / utils.BASE_RESOLUTION))
-        for molecule in self:
-            tempAxis += molecule.lineSurvey
-        return tempAxis
-
-    @property
-    def yAxis(self):
-        return np.zeros(int((self.rangeMax - self.rangeMin) / self.resolution))
-
-    @property
-    def xAxis(self):
-        return np.linspace(self.rangeMin, self.rangeMax, (self.rangeMax - self.rangeMin) / utils.BASE_RESOLUTION,
-                           endpoint=True)
-
-    @property
-    def absCoef(self):
-        if not self.progressAbsCoef:
-            for molecule in self:
-                self.absorptionCoefficient += getAbsCoef(molecule)
-        self.progressAbsCoef = True
-        return self.absorptionCoefficient
-
-    @property
-    def transmittance(self):
-        return np.exp(-self.absCoef * self.depth)
-
-    @property
-    def absorbance(self):
-        return np.log10(1 / self.transmittance)
-
-    @property
-    def title(self):
-        return '%s\nP: %smBars; T: %sK; depth: %scm' % (str(self), self.P, self.T, self.depth)
-
-    @property
-    def emissivity(self):
-        return 1 - self.transmittance
-
-    @property
-    def emittance(self):
-        return self.emissivity
-
-    @property
-    def molarMass(self):
-        mass = 0
-        for mol in self:
-            mass += mol.molarMass * mol.concentration
-        return mass
-
-    @property
-    def density(self):
-        d = self.P * 100 / self.T / self.specGasConstant
-        return d
-
-    @property
-    def mass(self):
-        return self.density * self.depth / 100
-
-    @property
-    def specGasConstant(self):
-        con = R * 1000 / self.molarMass
-        return con
-
-    @property
-    def temperatureAtHeight(self):
-        return int(self.atmosphere.temperatureAtHeight(self.meanHeight))
-
-    @property
-    def pressureAtHeight(self):
-        return self.atmosphere.pressureAtHeight(self.meanHeight)
-
-    @property
-    def densityAtHeight(self):
-        return self.atmosphere.densityAtHeight(self.temperatureAtHeight, self.specGasConstant)
-
-    def changeRange(self, rangeMin, rangeMax):
-        self.rangeMin = rangeMin
-        self.rangeMax = rangeMax
-        self.effectiveRangeMax = self.rangeMax + self.distanceFromCenter
-        self.effectiveRangeMin = max(self.rangeMin - self.distanceFromCenter, 0)
-        resetData(self)
-
-    def changeTemperature(self, temperature):
-        self.T = temperature
-        resetCrossSection(self)
-
-    def changePressure(self, pressure):
-        self.P = pressure
-        self.distanceFromCenter = self.P / 1013.25 * 5
-        if not self.dynamicResolution:
-            self.resolution = utils.BASE_RESOLUTION
-        else:
-            self.resolution = max(10**int(np.log10((self.P / 1013.25))) * .01, utils.BASE_RESOLUTION)
-        resetData(self)
-
-    def changeDepth(self, depth):
-        self.depth = depth
-
-    def addMolecule(self, name, isotopeDepth=1, **abundance):
-        molecule = Molecule(name, self, isotopeDepth, **abundance)
-        self.append(molecule)
-        if totalConcentration(self) > 1:
-            print('**Warning : Concentrations exceed 1.')
-        molecule.getData()
-        return molecule
-
-    def returnCopy(self, name=None):
-        if not name:
-            name = self.atmosphere.nextLayerName()
-        newCopy = Layer(self.depth, self.T, self.P, self.rangeMin, self.rangeMax, height=self.height,
-                        atmosphere=self.atmosphere, name=name, dynamicResolution=self.dynamicResolution)
-        for molecule in self:
-            newCopy.addMolecule(molecule.name, molecule.isotopeDepth, concentration=molecule.concentration)
-        print('%s copied to %s' % (self.name, newCopy.name))
-        return newCopy
-
-    def returnMoleculeObjects(self):
-        moleculeList = []
-        for m in self:
-            moleculeList.append(m)
-        return moleculeList
-
-    def planck(self, temperature):
-        return pyradPlanck.planckWavenumber(self.xAxis, temperature)
-
-    def transmission(self, surfaceSpectrum):
-        transmitted = self.transmittance * surfaceSpectrum
-        emitted = self.emittance * self.planck(self.T)
-        return transmitted + emitted
-
-
-class Atmosphere(list):
-    def __init__(self, name, rangeMin=0, rangeMax=0, planet=None):
-        super().__init__(self)
-        self.name = name
-        self.rangeMin = rangeMin
-        self.rangeMax = rangeMax
-        self.planet = planet
-
-    def __str__(self):
-        return self.name
-
-    def __bool__(self):
-        return True
-
-    def addLayer(self, depth, T, P, rangeMin, rangeMax, name=None, dynamicResolution=True, height=0.0):
-        if not name:
-            name = self.nextLayerName()
-        newLayer = Layer(depth, T, P, rangeMin, rangeMax, atmosphere=self, name=name,
-                         dynamicResolution=dynamicResolution, height=height)
-        return newLayer
-
-    def nextLayerName(self):
-        return 'Layer %s' % (len(self) + 1)
-
-    def returnLayerNames(self):
-        tempList = []
-        for layer in self:
-            tempList.append(layer.name)
-        return tempList
-
-    def returnLayerObjects(self):
-        if len(self) == 0:
-            return False
-        tempList = []
-        for layer in self:
-            tempList.append(layer)
-        return tempList
-
-    def temperatureAtHeight(self, height):
-        ruleList = self.planet.returnApplicableRules(height, 'temperature')
-        if len(ruleList) > 1:
-            print('Multiple rules found for height %s: %s' % (height, ruleList))
-        rule = ruleList[0]
-        temperature = rule.rateFunction(rule.baseValue, rule.baseHeight, rule.rate, height)
-        return temperature
-
-    def pressureAtHeight(self, height):
-        height = height / 100
-        ruleList = self.planet.returnApplicableRules(height, 'temperature')
-        if len(ruleList) > 1:
-            print('Multiple rules found for height %s: %s' % (height, ruleList))
-        rule = ruleList[0]
-        temperature = self.temperatureAtHeight(height)
-        if rule.rate != 0:
-            pressure = rule.basePressure * (rule.baseValue / temperature) ** \
-                  (self.planet.gravity * self.planet.molarMass / R / rule.rate)
-        else:
-            pressure = rule.basePressure * \
-                       np.exp(-self.planet.gravity * self.planet.molarMass *
-                              (height - rule.baseHeight) / R / rule.baseValue)
-        return pressure
-
-    def compositionAtHeight(self, height, molecule):
-        ruleList = self.planet.returnApplicableRules(height, 'composition')
-        if not ruleList:
-            return molecule.concentration
-        for rule in ruleList:
-            if rule.molecule.name == molecule.name:
-                concentration = rule.rateFunction(rule.baseValue, rule.baseHeight, rule.rate, height)
-                return concentration
-        return molecule.concentration
-
-    def densityAtHeight(self, height, gasConstant):
-        temperature = self.temperatureAtHeight(height)
-        pressure = self.pressureAtHeight(height)
-        density = pressure * 100 / gasConstant / temperature
-        return density
-
-
-class Planet:
-    def __init__(self, name, pressure, temperature, maxHeight, molarMass=0.0289644,
-                 gravity=9.80665, rangeMin=0, rangeMax=2000, initialThickness=100):
-        self.name = name
-        self.mass = 0
-        self.gravity = gravity
-        self.maxHeight = maxHeight * 100000
-        self.molarMass = molarMass
-        self.radius = 0
-        self.surfacePressure = pressure
-        self.surfaceTemperature = temperature
-        self.atmosphereRules = []
-        self.heightList = []
-        self.depthList = []
-        self.rangeMin = rangeMin
-        self.rangeMax = rangeMax
-        self.atmosphere = Atmosphere("%s's atmosphere" % self.name, planet=self)
-        self.initialLayer =  \
-            self.atmosphere.addLayer(initialThickness * 100, temperature, pressure, rangeMin, rangeMax,
-                                     name='initial layer', height=0)
-        self.progressProfileLoaded = False
-
-    def setMass(self, mass):
-        self.mass = mass
-
-    def addLapseRate(self, name, finalHeight, finalValue, rateFunction=linear):
-        self.atmosphereRules.append(AtmosphereRule(name, finalHeight * 100000,
-                                                   finalValue, 'temperature', self, rateFunction=rateFunction))
-
-    def addCompositionRate(self, name, finalHeight, finalValue, molecule, rateFunction=linear):
-        self.atmosphereRules.append(AtmosphereRule(name, finalHeight * 100000, finalValue, 'composition', self,
-            molecule=molecule, rateFunction=rateFunction))
-
-    def setRadius(self, radius):
-        self.radius = radius
-
-    def setGravity(self, gravity):
-        self.gravity = gravity
-
-    def setSurfacePressure(self, pressure):
-        self.surfacePressure = pressure
-
-    def setSurfaceTemperature(self, temperature):
-        self.surfaceTemperature = temperature
-
-    def setSurfaceComposition(self):
-        pass
-
-    def returnApplicableRules(self, height, ruleType):
-        ruleList = []
-        for rule in self.atmosphereRules:
-            if rule.isInRange(height, ruleType):
-                ruleList.append(rule)
-        if not ruleList:
-            return False
-        return ruleList
-
-    def returnAllRulesOfType(self, ruleType):
-        ruleList = []
-        for rule in self.atmosphereRules:
-            if rule.ruleType == ruleType:
-                ruleList.append(rule)
-        return ruleList
-
-    def temperatureAtHeight(self, height):
-        return self.atmosphere.temperatureAtHeight(height)
-
-    def pressureAtHeight(self, height):
-        return self.atmosphere.pressureAtHeight(height)
-
-    def densityAtHeight(self, height):
-        return self.atmosphere.densityAtHeight(height)
-
-    def compositionAtHeight(self, height, molecule):
-        return self.atmosphere.compositionAtHeight(height, molecule)
-
-    def addMolecule(self, name, isotopeDepth=1, **abundance):
-        molecule = self.initialLayer.addMolecule(name, isotopeDepth, **abundance)
-        return molecule
-
-    def sliceAtm(self, verify=True):
-        acceptSetup = False
-        layer = self.initialLayer
-        initialTemp = layer.T
-        initialPressure = layer.P
-        initialHeight = layer.height
-        initialDepth = layer.depth
-        while not acceptSetup:
-            mass = self.initialLayer.mass
-            print('%s, p %s, T %s, depth %s, height %s' % (layer.name, initialPressure, initialTemp, initialDepth, initialHeight))
-            self.heightList = [layer.height]
-            tempList = [layer.T]
-            pressureList = [layer.P]
-            self.depthList = [layer.depth]
-            while layer.height + layer.depth < self.maxHeight:
-                print('%s: K: %s, mBar: %s, baseHeight: %skm, depth: %skm, mass: %skg' % (
-                       'Layer %s' %len(self.heightList), int(layer.T), round(layer.P, 2), round(layer.height / 100000, 2), round(layer.depth / 100000, 2), round(layer.mass, 2)))
-                newHeight = self.heightList[-1] + self.depthList[-1]
-                tempList.append(self.temperatureAtHeight(newHeight))
-                pressureList.append(self.pressureAtHeight(newHeight))
-                layer.P = pressureList[-1]
-                layer.T = tempList[-1]
-                newDepth = (mass / layer.density * 100)
-                if newHeight + newDepth > self.maxHeight:
-                    newDepth = self.maxHeight - newHeight
-                    self.heightList.append(newHeight)
-                    self.depthList.append(newDepth)
-                    print('%s: K: %s, mBar: %s, baseHeight: %skm, depth: %skm, mass: %skg' % (
-                        'Layer %s' % len(self.heightList), int(layer.T), round(layer.P, 2),
-                        round(newHeight, 2), round(newDepth, 2), round(layer.mass, 2)))
-                    break
-                else:
-                    self.heightList.append(newHeight)
-                    self.depthList.append(newDepth)
-                    layer.height = newHeight
-                    layer.depth = newDepth
-            '''if self.heightList[-1] + .5 * self.depthList[-1] > self.maxHeight:
-                self.heightList.pop()
-                self.depthList.pop()'''
-            print('Total # of layers: %s' % len(self.heightList))
-            print('Total # of absorption lines per layer: %s' % len(totalLineList(layer)))
-            if verify:
-                if yesOrNo('Accept the current slicing (y/n): '):
-                    acceptSetup = True
-                else:
-                    validNumber = False
-                    while not validNumber:
-                        print(
-                            'Enter the new initial depth. Larger depth will decrease number of layers, smaller will increase')
-                        userNumber = input('Current depth is %s:' % utils.limeText('%sm' % (int(self.depthList[0]) / 100)))
-                        try:
-                            newDepth = float(userNumber)
-                            self.initialLayer.depth = newDepth * 100
-                            self.initialLayer.P = initialPressure
-                            self.initialLayer.T = initialTemp
-                            self.initialLayer.height = initialHeight
-                            validNumber = True
-                        except ValueError:
-                            print('Invalid number.')
-            else:
-                acceptSetup = True
-        return
-
-    def processLayers(self, verify=True):
-        if utils.profileProgress(self.name):
-            self.sliceAtm(verify=False)
-        if not self.heightList:
-            self.sliceAtm(verify=verify)
-        layer = self.initialLayer
-        startPoint = utils.profileProgress(self.name)
-        i = startPoint + 1
-        print('Incomplete profile found, resuming...')
-        for height, depth in zip(self.heightList[startPoint:], self.depthList[startPoint:]):
-            layer.name = 'Layer %s:%s' % (i, len(self.heightList))
-            layer.height = height
-            layer.depth = depth
-            layer.T = int(self.temperatureAtHeight(layer.meanHeight))
-            layer.P = self.pressureAtHeight(layer.meanHeight)
-            for molecule in layer:
-                molecule.concentration = self.compositionAtHeight(layer.meanHeight, molecule)
-            layer.createCrossSection()
-            utils.writePlanetProfile(self.name, layer)
-            utils.profileWriteProgress(self.name, i, len(self.heightList))
-            resetCrossSection(layer)
-            layer.absorptionCoefficient = np.zeros(len(layer.crossSection))
-            layer.progressAbsCoef = False
-            i += 1
-        utils.profileWriteComplete(self.name, i, len(self.heightList))
-        return
-
-    def loadProfile(self, verify=True):
-        if utils.profileComplete(self.name):
-            fileLength = utils.profileLength(self.name)
-            if verify:
-                if not yesOrNo('%s layers found for profile %s. Load profile (y/n):' % (fileLength, self.name)):
-                    if yesOrNo('Are you sure, selecting yes will delete all previous profile data (y/n): '):
-                        utils.emptyProfileDirectory(self.name)
-                        self.processLayers()
-        else:
-            self.processLayers()
-        fileLength = utils.profileLength(self.name)
-        while len(self.atmosphere) > 0:
-            self.atmosphere.pop()
-        for i in range(1, fileLength + 1):
-            lP = utils.readPlanetProfile(self.name, i, fileLength)
-            layer = Layer(lP['depth'], lP['T'], lP['P'], lP['rangeMin'], lP['rangeMax'], height=lP['height'], name=lP['name'])
-            self.atmosphere.append(layer)
-            layer.absorptionCoefficient = np.asarray(lP['absCoef'])
-            layer.progressAbsCoef = True
-        print('')
-        self.progressProfileLoaded = True
-        return
-
-    def processTransmission(self, height, direction='down', verify=True):
-        if not self.progressProfileLoaded:
-            self.loadProfile(verify=verify)
-        height = height * 100000
-        print('Processing atmosphere spectrum from %skm looking %s...' % (height / 100000, direction))
-        if direction == 'down':
-            surfaceSpectrum = pyradPlanck.planckWavenumber(self.initialLayer.xAxis, self.surfaceTemperature)
-            for layer in self.atmosphere:
-                if layer.meanHeight < height:
-                    surfaceSpectrum = layer.transmission(surfaceSpectrum)
-                    print('Processing %s...' % (layer.name), end='\r', flush=True)
-        elif direction == 'up':
-            surfaceSpectrum = pyradPlanck.planckWavenumber(self.initialLayer.xAxis, 3)
-            for layer in reversed(self.atmosphere):
-                if layer.meanHeight > height:
-                    surfaceSpectrum = layer.transmission(surfaceSpectrum)
-                    print('Processing %s...\n' % (layer.name), layer.T, end='\r', flush=True)
-        print('')
-        print('total power of spectrum: %sWm-2' % int((integrateSpectrum(surfaceSpectrum) / 5.67E-8)**.25))
-        return surfaceSpectrum
-
-    @property
-    def yAxis(self):
-        return np.zeros(int((self.rangeMax - self.rangeMin) / utils.BASE_RESOLUTION))
-
-    @property
-    def xAxis(self):
-        return np.linspace(self.rangeMin, self.rangeMax, (self.rangeMax - self.rangeMin) / utils.BASE_RESOLUTION,
-                           endpoint=True)
-
-    @property
-    def atmChangePoints(self):
-        changePointList = []
-        for rule in self.atmosphereRules:
-            changePointList.append(rule.finalHeight)
-        return sorted(changePointList)
-
-    @property
-    def transmission(self):
-        return self.processTransmission(self.maxHeight)
-
-
-def stefanB(power):
-    return (power / sb) ^ .25
-
-
-class AtmosphereRule:
-    def __init__(self, name, finalHeight, finalValue, ruleType, planetClass, molecule=None, rateFunction=linear):
-        self.name = name
-        self.ruleType = ruleType
-        self.parent = planetClass
-        self.baseHeight, self.baseValue, self.basePressure = self.setBaseValues(self.ruleType, molecule)
-        self.finalHeight = finalHeight
-        self.finalValue = finalValue
-        if rateFunction == linear:
-            self.rate = (self.finalValue - self.baseValue) / (self.finalHeight - self.baseHeight)
-        self.molecule = molecule
-        self.rateFunction = rateFunction
-
-    def isInRange(self, height, ruleType):
-        return self.baseHeight < height <= self.finalHeight and self.ruleType == ruleType
-
-    def baseHeightTemperatureRule(self):
-        height = 0
-        temperature = self.parent.surfaceTemperature
-        pressure = self.parent.surfacePressure
-        tempRules = self.parent.returnAllRulesOfType(self.ruleType)
-        for rule in tempRules:
-            if rule.finalHeight > height:
-                height = rule.finalHeight
-                temperature = rule.finalValue
-                pressure = self.parent.pressureAtHeight(height)
-        return height, temperature, pressure
-
-    def baseHeightCompositionRule(self, molecule):
-        height = 0
-        pressure = self.parent.surfacePressure
-        concentration = molecule.concentration
-        compRules = self.parent.returnAllRulesOfType(self.ruleType)
-        for rule in compRules:
-            if rule.finalHeight > height and rule.molecule == molecule:
-                height = rule.finalHeight
-                concentration = rule.finalValue
-        return height, concentration, pressure
-
-    def setBaseValues(self, ruleType, molecule):
-        if ruleType == 'temperature':
-            return self.baseHeightTemperatureRule()
-        elif ruleType == 'composition':
-            return self.baseHeightCompositionRule(molecule)
-
-
-def returnPlot(obj, propertyToPlot):
-    if propertyToPlot == "transmittance":
-        yAxis = getTransmittance(obj), 1
-    elif propertyToPlot == 'absorption coefficient':
-        yAxis = getAbsCoef(obj), 0
-    elif propertyToPlot == 'cross section':
-        yAxis = getCrossSection(obj), 0
-    elif propertyToPlot == 'absorbance':
-        yAxis = getAbsorbance(obj), 0
-    elif propertyToPlot == 'optical depth':
-        yAxis = getOpticalDepth(obj), 0
-    elif propertyToPlot == 'line survey':
-        yAxis = obj.lineSurvey, 0
-    else:
+def validPressure(userInput):
+    if not userInput:
         return False
-    return yAxis
+    try:
+        splitInput = validValueAndUnits.match(userInput)
+    except AttributeError:
+        print('Invalid input for pressure. Example: %s. %s'
+              % (util.limeText('1.35atm'), util.underlineMagenta('Please try again.')))
+        return False
+    unit = splitInput.group(5)
+    if not unit:
+        unit = 'mbar'
+    unit = unit.lower()
+    if unit not in PRESSURE_UNITS:
+        print('Invalid units. Accepted units are %s.' % ', '.join(PRESSURE_UNITS))
+        return False
+    textNumber = ''
+    for i in range(1, 5):
+        if splitInput.group(i):
+            textNumber += splitInput.group(i)
+
+    value = float(textNumber)
+    return pyradClasses.convertPressure(value, unit)
 
 
-def isBetween(test, minValue, maxValue):
-    if test >= minValue:
-        if test <= maxValue:
-            return True
-    return False
+def validComposition(userInput):
+    if not userInput:
+        return False
+    try:
+        splitInput = validValueAndUnits.match(userInput)
+    except AttributeError:
+        print('Invalid input for concentration. Example: %s. %s'
+              % (util.limeText('15ppb'), util.underlineMagenta('Please try again.')))
+        return False
+    unit = splitInput.group(5)
+    if not unit:
+        unit = 'concentration'
+    if unit not in COMPOSITION_UNITS:
+        print('Invalid units. Accepted units are %s.'
+              % (util.limeText(', '.join(COMPOSITION_UNITS))))
+        return False
+    textNumber = ''
+    for i in range(1, 5):
+        if splitInput.group(i):
+            textNumber += splitInput.group(i)
+    value = float(textNumber)
+    if value <= 0:
+        print('Concentration must be greater than 0')
+        return False
+    return value, unit
 
 
-def plot(propertyToPlot, title, plotList, fill=False):
-    plt.figure(figsize=(10, 6), dpi=80)
-    plt.subplot(111, facecolor='xkcd:dark grey')
-    plt.xlabel('wavenumber cm-1')
-    plt.margins(0.01)
-    plt.subplots_adjust(left=.07, bottom=.08, right=.97, top=.90)
-    plt.ylabel(propertyToPlot)
-    if propertyToPlot == 'line survey':
-        plt.yscale('log')
-    plt.grid('grey', linewidth=.5, linestyle=':')
-    plt.title('%s' % title)
-    handles = []
-    linewidth = 1.2
-    alpha = .7
-    for singlePlot, color in zip(plotList, COLOR_LIST):
-        yAxis, fillAxis = returnPlot(singlePlot, propertyToPlot)
-        fig, = plt.plot(singlePlot.xAxis,  smooth(yAxis, 50), linewidth=linewidth, alpha=alpha, color=color,
-                        label='%s' % singlePlot.name)
-        handles.append(fig)
-        plt.fill_between(singlePlot.xAxis, fillAxis, yAxis, color=color, alpha=.3 * fill)
-        linewidth = .7
-        alpha = .5
-    legend = plt.legend(handles=handles, frameon=False)
-    text = legend.get_texts()
-    plt.setp(text, color='w')
-    plt.show()
+def validTemperature(userInput):
+    if not userInput:
+        return False
+    try:
+        splitInput = validValueAndUnits.match(userInput)
+    except AttributeError:
+        print('Invalid input for temperature. Example: %s. %s'
+              % (util.limeText('20C'), util.underlineMagenta('Please try again.')))
+        return False
+    unit = splitInput.group(5)
+    if not unit:
+        unit = 'K'
+    unit = unit.upper()[0]
+    if unit not in TEMPERATURE_UNITS:
+        print('Invalid units. Accepted units are %s.'
+              % (util.limeText(', '.join(TEMPERATURE_UNITS))))
+        return False
+    textNumber = ''
+    for i in range(1, 5):
+        if splitInput.group(i):
+            textNumber += splitInput.group(i)
+    value = float(textNumber)
+    return pyradClasses.convertTemperature(value, unit)
 
 
-def plotSpectrum(layer=None, title=None, rangeMin=None, rangeMax=None, objList=None, surfaceSpectrum=None,
-                 planckTemperatureList=None, planckType='wavenumber'):
-    plt.figure(figsize=(10, 6), dpi=80)
-    plt.subplot(111, facecolor='xkcd:dark grey')
-    plt.margins(0.01)
-    plt.subplots_adjust(left=.07, bottom=.08, right=.97, top=.90)
-    if layer:
-        rangeMin = layer.rangeMin
-        rangeMax = layer.rangeMax
-        title = layer.title
-    if planckType == 'wavenumber':
-        plt.xlabel('wavenumber cm-1')
-        plt.ylabel('Radiance Wm-2sr-1(cm-1)-1')
-        planckFunction = pyradPlanck.planckWavenumber
-        xAxis = np.linspace(rangeMin, rangeMax, (rangeMax - rangeMin) / utils.BASE_RESOLUTION)
-    elif planckType == 'Hz':
-        plt.xlabel('Hertz')
-        plt.ylabel('Radiance Wm-2sr-1Hz-1')
-        planckFunction = pyradPlanck.planckHz
-        xAxis = np.linspace(rangeMin, rangeMax, 1000)
-    else:
-        plt.xlabel('wavelength um')
-        plt.ylabel('Radiance Wm-2sr-1um-1')
-        planckFunction = pyradPlanck.planckWavelength
-        xAxis = np.linspace(rangeMin, rangeMax, (rangeMax - rangeMin) / utils.BASE_RESOLUTION)
-    plt.title('%s' % title)
-    handles = []
-    blue = .3
-    red = 1
-    green = .6
-    dr = -.15
-    db = .15
-    dg = .15
-    if not rangeMax:
-        xAxis = layer.xAxis
-    for temperature in planckTemperatureList:
-        yAxis = planckFunction(xAxis, float(temperature))
-        fig, = plt.plot(xAxis, yAxis, linewidth=.75, color=(red, green, blue),
-                        linestyle=':', label='%sK : %sWm-2' %
-                        (temperature, round(integrateSpectrum(yAxis, res=(rangeMax - rangeMin) / len(yAxis)), 2)))
-        handles.append(fig)
-        if red + dr < 0 or red + dr > 1:
-            dr *= -1
-        if green + dg > 1 or green + dg < 0:
-            dg *= -1
-        if blue + db > 1 or blue + db < 0:
-            db *= -1
-        red += dr
-        green += dg
-        blue += db
-        if red < .3 and green < .3 and blue < .3:
-            green += .5
-            blue += .2
-        if red < .3 and green < .3:
-            green += .4
-    if objList:
-        alpha = .7
-        linewidth = 1.2
-        for obj, color in zip(objList, COLOR_LIST):
-            yAxis = obj.transmission(surfaceSpectrum)
-            fig, = plt.plot(layer.xAxis, yAxis, linewidth=linewidth,
-                            alpha=alpha, color=color,
-                            label='%s : %sWm-2' % (obj.name, round(integrateSpectrum(yAxis, pi), 2)))
-            handles.append(fig)
-            alpha = .5
-            linewidth = 1
-    legend = plt.legend(handles=handles, frameon=False)
-    text = legend.get_texts()
-    plt.setp(text, color='w')
-    plt.show()
+def validRange(userInput):
+    if not userInput:
+        return False
+    try:
+        splitInput = validValueAndUnits.match(userInput)
+    except AttributeError:
+        print('Invalid input for range. Example: %s. %s'
+              % (util.limeText('150cm-1'), util.underlineMagenta('Please try again.')))
+        return False
+    unit = splitInput.group(5)
+    if not unit:
+        unit = 'cm-1'
+    unit = unit.lower()
+    if unit not in RANGE_UNITS:
+        print('Invalid units. Accepted units are %s'
+              % (util.limeText(', '.join(RANGE_UNITS))))
+        return False
+    textNumber = ''
+    for i in range(1, 5):
+        if splitInput.group(i):
+            textNumber += splitInput.group(i)
+    value = float(textNumber)
+    return pyradClasses.convertRange(value, unit)
 
 
-def plotPlanetSpectrum(planets, height=None, direction='down', temperatureList=[300, 280, 250, 210, 170], verify=True, integrateRange=[]):
-    linewidth = .5
-    alpha = .7
-    plt.figure(figsize=(10, 6), dpi=80)
-    plt.subplot(111, facecolor='xkcd:almost black')
-    plt.margins(0.01)
-    plt.subplots_adjust(left=.07, bottom=.08, right=.97, top=.90)
-    plt.ylabel('Radiance Wm-2sr-1(cm-1)-1')
-    handles = []
-    heightFlag = True
-    if height is None:
-        heightFlag = False
-    for planet, color in zip(planets, COLOR_LIST[1:]):
-        if not heightFlag:
-            height = planet.maxHeight / 100000
-        xAxis = planet.xAxis
-        #xAxis = np.linspace(planet.rangeMin, planet.rangeMax, planet.rangeMax - planet.rangeMin)
-        yAxis = planet.processTransmission(height, direction=direction, verify=verify)
-        #yAxis = reduceRes(planet.processTransmission(height, direction=direction, verify=verify))
-        fig, = plt.plot(xAxis, smooth(yAxis, 75), linewidth=linewidth,
-                        alpha=alpha, color=color,
-                        label='%s : %sWm-2' % (planet.name, round(integrateSpectrum(yAxis, pi), 2)))
-        #fig, = plt.plot(xAxis, yAxis, linewidth=linewidth,
-        #                alpha=alpha, color=color,
-        #                label='%s : %sWm-2' % (planet.name, round(integrateSpectrum(yAxis, pi), 2)))
-        handles.append(fig)
-    color = 1
-    for temperature in temperatureList:
-        yAxis = pyradPlanck.planckWavenumber(xAxis, float(temperature))
-        fig, = plt.plot(xAxis, yAxis, linewidth=.75, color=(color, color, color), alpha=alpha,
-                        linestyle=':', label='%sK : %sWm-2' %
-                                             (temperature,
-                                              round(integrateSpectrum(yAxis, pi), 2)))
-        handles.append(fig)
-        color -= .15
-    legend = plt.legend(handles=handles, frameon=False)
-    text = legend.get_texts()
-    plt.setp(text, color='w')
-    plt.show()
+def validDepth(userInput):
+    if not userInput:
+        return False
+    try:
+        splitInput = validValueAndUnits.match(userInput)
+    except AttributeError:
+        print('Invalid input for depth. Example: %s. %s'
+              % (util.limeText('10cm'), util.underlineMagenta('Please try again.')))
+        return False
+    unit = splitInput.group(5)
+    if not unit:
+        unit = 'cm'
+    unit = unit.lower()
+    if unit not in DEPTH_UNITS:
+        print('Invalid units. Accepted units are %s'
+              % (util.limeText(', '.join(DEPTH_UNITS))))
+        return False
+    textNumber = ''
+    for i in range(1, 5):
+        if splitInput.group(i):
+            textNumber += splitInput.group(i)
+    value = float(textNumber)
+    return pyradClasses.convertLength(value, unit)
 
 
-HITRAN_GLOBAL_ISO = {1: {1: 1, 2: 2, 3: 3, 4: 4, 5: 5, 6: 6, 7: 129},
-                     2: {1: 7, 2: 8, 3: 9, 4: 10, 5: 11, 6: 12, 7: 13, 8: 14, 9: 121, 10: 15, 11: 120, 12: 122},
-                     3: {1: 16, 2: 17, 3: 18, 4: 19, 5: 20},
-                     4: {1: 21, 2: 22, 3: 23, 4: 24, 5: 25, },
-                     5: {1: 26, 2: 27, 3: 28, 4: 29, 5: 30, 6: 31},
-                     6: {1: 32, 2: 33, 3: 34, 4: 35},
-                     7: {1: 36, 2: 37, 3: 38},
-                     8: {1: 39, 2: 40, 3: 41},
-                     9: {1: 42, 2: 43},
-                     10: {1: 44},
-                     11: {1: 45, 2: 46},
-                     12: {1: 47, 2: 117},
-                     13: {1: 48, 2: 49, 3: 50},
-                     14: {1: 51, 2: 110},
-                     15: {1: 52, 2: 53, 3: 107, 4: 108},
-                     16: {1: 19, 2: 11, 3: 111, 4: 112},
-                     17: {1: 56, 2: 113},
-                     18: {1: 57, 2: 58},
-                     19: {1: 59, 2: 60, 3: 61, 4: 62, 5: 63},
-                     20: {1: 64, 2: 65, 3: 66},
-                     21: {1: 67, 2: 68},
-                     22: {1: 69, 2: 118},
-                     23: {1: 70, 2: 71, 3: 72},
-                     24: {1: 73, 2: 74},
-                     25: {1: 75},
-                     26: {1: 76, 2: 77, 3: 105},
-                     27: {1: 78, 2: 106},
-                     28: {1: 79},
-                     29: {1: 80, 2: 119},
-                     30: {1: 126},
-                     31: {1: 81, 2: 82, 3: 83},
-                     32: {1: 84},
-                     33: {1: 85},
-                     34: {1: 86},
-                     35: {1: 127, 2: 128},
-                     36: {1: 87},
-                     37: {1: 88, 2: 89},
-                     38: {1: 90, 2: 91},
-                     39: {1: 92},
-                     40: {1: 93, 2: 94},
-                     41: {1: 95},
-                     42: {1: 96},
-                     43: {1: 116},
-                     44: {1: 109},
-                     45: {1: 103, 2: 115},
-                     46: {1: 97, 2: 98, 3: 99, 4: 100},
-                     47: {1: 114},
-                     48: {1: 123},
-                     49: {1: 124, 2: 125},
-                     901: {1: 901}}
+DEPTH_UNITS = ['cm', 'in', 'inches', 'ft', 'feet', 'meter', 'm']
+PRESSURE_UNITS = ['atm', 'bar', 'mbar', 'pa']
+TEMPERATURE_UNITS = ['K', 'C', 'F']
+RANGE_UNITS = ['um', 'cm-1']
+COMPOSITION_UNITS = ['ppm', 'ppb', '%', 'percentage', 'perc', 'concentration']
 
-COLOR_LIST = ['xkcd:white',
-              'xkcd:bright orange',
-              'xkcd:seafoam green',
-              'xkcd:bright blue',
-              'xkcd:salmon',
-              'xkcd:light violet',
-              'xkcd:green yellow']
-
-VERSION = utils.VERSION
-
-MOLECULE_ID = {'h2o': 1, 'co2': 2, 'o3': 3, 'n2o': 4, 'co': 5,
-               'ch4': 6, 'o2': 7, 'no': 8, 'so2': 9,
-               'no2': 10, 'nh3': 11, 'hno3': 12, 'oh': 13,
-               'hf': 14, 'hcl': 15, 'hbr': 16, 'hi': 17,
-               'clo': 18, 'ocs': 19, 'h2co': 20, 'hocl': 21,
-               'n2': 22, 'hcn': 23, 'ch3cl': 24, 'h2o2': 25,
-               'c2h2': 26, 'c2h6': 27, 'ph3': 28, 'cof2': 29,
-               'sf6': 30, 'h2s': 31, 'hcooh': 32, 'ho2': 33,
-               'o': 34, 'clono2': 35, 'no+': 36, 'hobr': 37,
-               'c2h4': 38, 'ch3oh': 39, 'ch3br': 40, 'ch3cn': 41,
-               'cf4': 42, 'c4h2': 43, 'hc3n': 44, 'h2': 45,
-               'cs': 46, 'so3': 47, 'c2n2': 48, 'cocl2': 49,
-               'ar': 901}
-
-
-if __name__ == 'main':
-    pyradInteractive.menuMain()
+while True:
+    menuMain()
